Marking off invalid calendar days
