--- app/components/product-details/daterange.component.css
@@ -1,7 +1,0 @@
-main {
-    padding: 1em;
-    font-family: Arial, Helvetica, sans-serif;
-    margin-top: 50px;
-    display: block;
-    max-width: 600px;
-}
