--- app/components/product-details/daterange.component.html
@@ -1,1 +1,0 @@
-<input class="form-control" type="text" name="daterangeInput" daterangepicker [options]="options" (selected)="selectedDate($event)" />
