+++ app/components/product-details/product-details.ts
@@ -55,21 +55,20 @@ export class ProductDetails implements OnInit {
     private uiRouter: UIRouter,
     private drpOptions: DaterangepickerConfig,
   ) {
+    // get invalid dates from transaction table
     this.prodId = this.uiRouter.globals.params["productId"];
     this.productDetailsService
     .getInvalidDays(this.prodId)
     .then(response => {
       const invalidDates = response;
       this.invalidDays = invalidDates;
-      console.log(this.invalidDays);
-      // format invalid days
 
+    // format invalid days
       let oneDay = 1000 * 60 * 60 * 24;
 
-    // days.forEach(function(day: any) {
       for(var i = 0; i < this.invalidDays.length; i++) {
-        let from = moment(this.invalidDays[i].bookedfrom);
-        let to = moment(this.invalidDays[i].bookedto);
+        let from: any = moment(this.invalidDays[i].bookedfrom);
+        let to: any = moment(this.invalidDays[i].bookedto);
         let diffMs = to - from;
         let numDays = Math.round(diffMs / oneDay);
 
@@ -81,53 +80,26 @@ export class ProductDetails implements OnInit {
           numDays--;
         } while (numDays > 0);
       }
-
-      // console.log("this Sucks", this, this.formattedDays);
-
-      // this.drpOptions.settings = {
-      //   opens: "center",
-      //   isInvalidDate: function(date: any) {
-      //   for(var i = 0; i < this.formattedDays.length; i++) {
-      //     if(date.format('MM-DD-YYYY') == this.formattedDays[i]) {
-      //       return true;
-      //     } else {
-      //       return false;
-      //       }
-      //     }
-      //   }
-      // }
     })
-
-    console.log('outside promise', this, this.formattedDays);
 
     config.max = 5;
     config.readonly = true;
 
-    // this.drpOptions.settings = {
-    //   opens: "center",
-    //   isInvalidDate: function(date: any) {
-    //     console.log(date);
-    //     if(date.format('MM-DD-YYYY') == '11-30-2016') {
-    //       return true;
-    //     } else {
-    //       return false;
-    //     }
-    //   }
+    let context = this;
+
     this.drpOptions.settings = {
       opens: "center",
       isInvalidDate: function(date: any) {
-        console.log('running invalid dates', this);
-        for(var i = 0; i < this.formattedDays.length; i++) {
-          if(date.format('MM-DD-YYYY') == this.formattedDays[i]) {
+        for(var i = 0; i < context.formattedDays.length; i++) {
+          if(date.format('MM-DD-YYYY') == context.formattedDays[i]) {
             return true;
-          } else {
-            return false;
           }
         }
       }
     }
   }
 
+// Product Details Methods
 
   public selectedDate(value: any) {
     this.fromDate = value.start;
@@ -148,102 +120,11 @@ export class ProductDetails implements OnInit {
     this.product = this.product[0];
     this.selectedPic = this.product.url[0];
     this.getReviews(this.product.id);
-    // this.getInvalidDays(this.product.id);
-    // this.drpOptions.settings.opens = "left";
-    // console.log(this.drpOptions.settings.isInvalidDate);
   }
 
   public onSelect(n: number) {
     this.selectedPic = this.product.url[n];
   }
-
-  // public getInvalidDays(productId: number) {
-  //   this.productDetailsService
-  //   .getInvalidDays(productId)
-  //   .then(response => {
-  //     const invalidDates = response;
-  //     this.invalidDays = invalidDates;
-  //     console.log(this.invalidDays);
-  //     // format invalid days
-  //     this.formatInvalidDays(this.invalidDays);
-  //     // this.drpOptions.settings.isInvalidDate = function(date: any) {
-  //     //   console.log('running invalid dates');
-  //     //   for(var i = 0; i < this.formattedDays.length; i++) {
-  //     //     if(date.format('MM-DD-YYYY') == this.formattedDays[i]) {
-  //     //       return true;
-  //     //     } else {
-  //     //       return false;
-  //     //     }
-  //     //   }
-  //     // }
-  //     // this.drpOptions.settings.isInvalidDate();
-  //   })
-  // }
-
-  // public formatInvalidDays(days: Array<any>) {
-
-  //   let oneDay = 1000 * 60 * 60 * 24;
-  //   console.log(days.length);
-
-  //   // days.forEach(function(day: any) {
-  //   for(var i = 0; i < days.length; i++) {
-  //     console.log(days);
-  //     // console.log(day.bookedto.slice(0, 10));
-  //     let from = moment(days[i].bookedfrom);
-  //     let to = moment(days[i].bookedto);
-  //     let diffMs = to - from;
-  //     let numDays = Math.round(diffMs / oneDay);
-
-  //     let currentDay = from;
-
-  //     do {
-  //       this.formattedDays.push(currentDay.format("MM-DD-YYYY"));
-  //       currentDay = currentDay.add(1, 'd');
-  //       numDays--;
-  //     } while (numDays > 0);
-
-  //     console.log(this.formattedDays);
-  //     let context = this;
-  //     // console.log(from.format("DD-MM-YYYY"));
-  //     // let nextDay = from.add(1, 'd');
-  //     // console.log(nextDay.format("DD-MM-YYYY"));
-
-  //     // returns milliseconds
-  //     // console.log(to - from);
-  //     // let to = (day.bookedto.slice(0, 10));
-  //     // let daysBetween =
-  //     // console.log(from, to);
-  //     // console.log(from + oneDay);
-  //     // let daysBetween = (to - from) / oneDay;
-  //     // console.log(daysBetween);
-  //     // while(daysBetween !== 0) {
-  //     // this.formattedDays.push(from, to);
-  //     // if(to.format('YYYY-MM-DD') == to) {
-  //     //   return true;
-  //     // }
-  //   }
-  //   // this.drpOptions.settings = {
-  //   //   opens: "center",
-  //   //   isInvalidDate: function(date: any) {
-  //   //     console.log('running invalid dates', context.formattedDays.length);
-  //   //     for(var i = 0; i < context.formattedDays.length; i++) {
-  //   //       if(date.format('MM-DD-YYYY') == context.formattedDays[i]) {
-  //   //         return true;
-  //   //       } else {
-  //   //         return false;
-  //   //       }
-  //   //     }
-  //   //   }
-  //   // }
-  //     // this.drpOptions.settings.isInvalidDate = function(date: any) {
-  //     //   if(date.format('MM-DD-YYYY') == '11-30-2016') {
-  //     //     return true;
-  //     //   } else {
-  //     //     return false;
-  //     //   }
-  //     // }
-  // // this.drpOptions.settings.isInvalidDate();    // this.drpOptions.settings.isInvalidDate();
-  // }
 
   public getReviews(productId: number) {
     this.productDetailsService
@@ -260,7 +141,6 @@ export class ProductDetails implements OnInit {
         console.log(this.reviews);
       })
       .catch(err => console.log(err));
-
   }
 
   public openCheckOut() {
@@ -285,7 +165,6 @@ export class ProductDetails implements OnInit {
       name: "Gear Box",
       amount: +this.totalAmount * 100,
     });
-
   }
 
   public convertObjToDate(obj: any) {
