+++ server/transactions/transactionController.js
@@ -55,5 +55,17 @@ module.exports = {
       }
       res.json(result.rows);
     })
+  },
+
+  getInvalidDays: function (req, res, next) {
+    var id = req.params.id;
+
+    var queryStr = `SELECT bookedfrom, bookedto FROM transactions WHERE product_id=($1)`;
+
+    pool.query(queryStr, [id], function(err, result) {
+      if (err) return console.log(err);
+      console.log('dates', result);
+      res.json(result.rows);
+    });
   }
 }
